Minor changes to Watermask file installation . l2gen_aquarius now reads in ProductInfo xml

+++ seadas-processing/src/main/java/gov/nasa/gsfc/seadas/processing/core/L2genData.java
@@ -21,7 +21,6 @@ import java.io.*;
 import java.net.URL;
 import java.util.*;
 import java.util.logging.Level;
-import java.util.regex.Pattern;
 
 /**
  * A ...
@@ -53,6 +52,11 @@ public class L2genData implements L2genDataProcessorModel {
             AQUARIUS_PARAM_CATEGORY_INFO_XML = "aquariusParamCategoryInfo.xml",
             AQUARIUS_PRODUCT_CATEGORY_INFO_XML = "aquariusProductCategoryInfo.xml";
 
+    private static final String
+            GETANC = "getanc",
+            GETANC_AQUARIUS = "getanc_aquarius.py";
+
+    private static File aquariusProductXmlFile;
 
     public static final String
             PAR = "par",
@@ -1137,8 +1141,20 @@ public class L2genData implements L2genDataProcessorModel {
         String ifile = getParamValue(getParamInfo(IFILE));
         StringBuilder ancillaryFiles = new StringBuilder("");
 
-        ProcessorModel processorModel = new ProcessorModel("getanc.py");
+        String getanc;
+
+        switch (mode) {
+            case L2GEN_AQUARIUS:
+                getanc = GETANC_AQUARIUS;
+                break;
+            default:
+                getanc = GETANC;
+                break;
+        }
+
+        ProcessorModel processorModel = new ProcessorModel(getanc);
         processorModel.setAcceptsParFile(false);
+
 
         int position = 1;
         if (refreshDB) {
@@ -1169,7 +1185,7 @@ public class L2genData implements L2genDataProcessorModel {
                 line = stdInput.readLine();
             }
         } catch (IOException e) {
-            System.out.println("ERROR - Problem running getanc.py");
+            System.out.println("ERROR - Problem running "+ getanc);
             System.out.println(e.getMessage());
             return;
         }
@@ -1316,35 +1332,27 @@ public class L2genData implements L2genDataProcessorModel {
         File ofile;
         ProcessorModel processorModel;
 
-        switch (mode) {
-            case L2GEN_AQUARIUS:
-                xmlFile = new File(l2genDir, AQUARIUS_PRODUCT_INFO_XML);
-                ofile = new File(l2genDir, AQUARIUS_PRODUCT_INFO_XML + ".out");
-                processorModel = new ProcessorModel("l2gen_aquarius");
-                break;
-            default:
-                xmlFile = new File(l2genDir, PRODUCT_INFO_XML);
-                ofile = new File(l2genDir, PRODUCT_INFO_XML + ".out");
-                processorModel = new ProcessorModel("l2gen");
-
-                break;
-        }
+        xmlFile = new File(l2genDir, PRODUCT_INFO_XML);
+        ofile = new File(l2genDir, PRODUCT_INFO_XML + ".out");
+        processorModel = new ProcessorModel("l2gen");
 
 
         processorModel.setAcceptsParFile(true);
 
         processorModel.addParamInfo("ifile", file.getAbsolutePath(), ParamInfo.Type.IFILE);
-   //     processorModel.addParamInfo("ofile", ofile.getAbsolutePath(), ParamInfo.Type.OFILE);
+        //     processorModel.addParamInfo("ofile", ofile.getAbsolutePath(), ParamInfo.Type.OFILE);
         processorModel.addParamInfo("prodxmlfile", xmlFile.getAbsolutePath(), ParamInfo.Type.OFILE);
 
         try {
-            Process p = OCSSWRunner.execute(processorModel.getProgramCmdArray(), processorModel.getIFileDir());//processorModel.executeProcess();
-            p.waitFor();
-            ofile.delete();
-            if (p.exitValue() != 0) {
-                System.out.println("ERROR - Problem creating Product XML file");
-                System.out.println("Exit value = " + Integer.toString(p.exitValue()));
-                return null;
+            if (mode != Mode.L2GEN_AQUARIUS) {
+                Process p = OCSSWRunner.execute(processorModel.getProgramCmdArray(), processorModel.getIFileDir());//processorModel.executeProcess();
+                p.waitFor();
+                ofile.delete();
+                if (p.exitValue() != 0) {
+                    System.out.println("ERROR - Problem creating Product XML file");
+                    System.out.println("Exit value = " + Integer.toString(p.exitValue()));
+                    return null;
+                }
             }
 
             if (!xmlFile.exists()) {
@@ -1363,6 +1371,26 @@ public class L2genData implements L2genDataProcessorModel {
         }
 
 //        return L2genForm.class.getResourceAsStream(PRODUCT_INFO_XML);
+    }
+
+    private InputStream getAquariusProductInfoInputStream() {
+        File file = getAquariusProductXmlFile();
+        File dataDir = SystemUtils.getApplicationDataDir();
+        File l2genDir = new File(dataDir, "l2gen");
+        l2genDir.mkdirs();
+
+        try {
+            if (file == null || !file.exists()) {
+                return null;
+            }
+
+            return new FileInputStream(file);
+        } catch (IOException e) {
+            System.out.println("ERROR - Problem creating Aquarius Product XML file");
+            System.out.println(e.getMessage());
+            return null;
+
+        }
     }
 
 
@@ -1513,7 +1541,18 @@ public class L2genData implements L2genDataProcessorModel {
     }
 
     public L2genProductsParamInfo createL2prodParamInfo(String value) {
-        InputStream productInfoStream = getProductInfoInputStream(getTinyIFile());
+
+        InputStream productInfoStream;
+
+        switch (mode) {
+            case L2GEN_AQUARIUS:
+                productInfoStream = getAquariusProductInfoInputStream();
+                break;
+            default:
+                productInfoStream = getProductInfoInputStream(getTinyIFile());
+                break;
+        }
+
 
         L2genProductsParamInfo l2prodParamInfo = new L2genProductsParamInfo();
         setL2prodParamInfo(l2prodParamInfo);
@@ -1540,19 +1579,23 @@ public class L2genData implements L2genDataProcessorModel {
         return StringUtils.join(productArrayList, " ");
     }
 
+
     public File getTinyIFile() {
         if (tinyIFile == null) {
-            tinyIFile = installTinyIFile();
+            tinyIFile = installResource(TINY_IFILE_NAME);
         }
         return tinyIFile;
     }
 
-    public static File installTinyIFile() {
-        return installTinyIFile(TINY_IFILE_NAME);
+    public File getAquariusProductXmlFile() {
+        if (aquariusProductXmlFile == null) {
+            aquariusProductXmlFile = installResource(AQUARIUS_PRODUCT_INFO_XML);
+        }
+        return aquariusProductXmlFile;
     }
 
-    public static File installTinyIFile(String fileName) {
-        final String tinyFileName = fileName;
+
+    public static File installResource(final String fileName) {
         final File dataDir = new File(SystemUtils.getApplicationDataDir(), "l2gen");
         File theFile = new File(dataDir, fileName);
         if (theFile.canRead()) {
@@ -1566,7 +1609,7 @@ public class L2genData implements L2genDataProcessorModel {
                 "Installing Auxdata...") {
             @Override
             protected Object doInBackground(ProgressMonitor progressMonitor) throws Exception {
-                resourceInstaller.install(tinyFileName, progressMonitor);
+                resourceInstaller.install(fileName, progressMonitor);
                 return Boolean.TRUE;
             }
